requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
astunparse==1.6.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Cython==0.29.21
#dataclasses==0.8
dnspython==1.16.0
fastapi==0.63.0
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
future==0.18.2
gdown==3.12.2
gunicorn==20.0.4
h11==0.11.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
logger==1.4
MarkupSafe==1.1.1
matplotlib==3.3.3
numpy==1.19.4
opencv-contrib-python==4.1.2.30
opencv-python==4.4.0.46
Pillow==8.0.1
pydantic==1.7.3
pymongo==3.11.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.25.1
scipy==1.5.4
six==1.15.0
starlette==0.13.6
#tensorflow-gpu-estimator==2.3.0
torch==1.7.1+cpu
torchaudio==0.7.2
torchvision==0.8.2+cpu
tqdm==4.54.1
typing-extensions==3.7.4.3
urllib3==1.26.2
uvicorn==0.13.2
Werkzeug==1.0.1
scikit_learn==0.21.3
numpy
matplotlib
torch
torchvision
terminaltables
pillow
tqdm
numba
fire
filterpy
opencv-python
scikit-image
PyQt5
pandas
wincertstore==0.2
scikit_learn==0.21.3
numpy
matplotlib
torch
torchvision
terminaltables
pillow
tqdm
numba
fire
filterpy
opencv-python
scikit-image
PyQt5
pandas
